Allow to sort Grid data without drag & drop option
The sortField is used for order by in SQL Statement. It was not possible
to use a order by value without the drag & drop sort option in a grid.
Now it's possible to order by name, or date fields. You have simple to
add 'allowSort' => "false" in Plugins "AdminController.php"